Revert "Move policy ingest service to a common service"
This reverts commit 2e1801c32ba813cf0342418d98cef590a971db8a.
