+++ x-pack/plugins/security_solution/public/management/pages/policy/store/services/ingest.test.ts
@@ -9,15 +9,9 @@ import {
   INGEST_API_EPM_PACKAGES,
   sendGetPackagePolicy,
   sendGetEndpointSecurityPackage,
-  sendGetEndpointSpecificPackagePolicies,
 } from './ingest';
 import { httpServiceMock } from '../../../../../../../../../src/core/public/mocks';
-import {
-  EPM_API_ROUTES,
-  PACKAGE_POLICY_SAVED_OBJECT_TYPE,
-  PACKAGE_POLICY_API_ROOT,
-  PACKAGE_POLICY_API_ROUTES,
-} from '../../../../../../../fleet/common';
+import { EPM_API_ROUTES, PACKAGE_POLICY_API_ROOT } from '../../../../../../../fleet/common';
 import { policyListApiPathHandlers } from '../test_mock_utils';
 
 describe('ingest service', () => {
@@ -25,29 +19,6 @@ describe('ingest service', () => {
 
   beforeEach(() => {
     http = httpServiceMock.createStartContract();
-  });
-
-  describe('sendGetEndpointSpecificPackagePolicies()', () => {
-    it('auto adds kuery to api request', async () => {
-      await sendGetEndpointSpecificPackagePolicies(http);
-      expect(http.get).toHaveBeenCalledWith(`${PACKAGE_POLICY_API_ROUTES.LIST_PATTERN}`, {
-        query: {
-          kuery: `${PACKAGE_POLICY_SAVED_OBJECT_TYPE}.package.name: endpoint`,
-        },
-      });
-    });
-    it('supports additional KQL to be defined on input for query params', async () => {
-      await sendGetEndpointSpecificPackagePolicies(http, {
-        query: { kuery: 'someValueHere', page: 1, perPage: 10 },
-      });
-      expect(http.get).toHaveBeenCalledWith(`${PACKAGE_POLICY_API_ROUTES.LIST_PATTERN}`, {
-        query: {
-          kuery: `someValueHere and ${PACKAGE_POLICY_SAVED_OBJECT_TYPE}.package.name: endpoint`,
-          perPage: 10,
-          page: 1,
-        },
-      });
-    });
   });
 
   describe('sendGetPackagePolicy()', () => {
